finish chapter5

+++ chapter5/src/c_5_4.cc
@@ -3,17 +3,18 @@ using namespace std;
 
 int main()
 {
- double Dephone = 100;
- double Cleo = 100;
+ float Dephone = 100;
+ float Cleo = 100;
  int year = 1;
 
- for(;Dephone >= Cleo;++year)
-  { 
-   Dephone += 10;
-   Cleo += (Cleo * 0.05);
-  } 
+  for(;Dephone >= Cleo; ++year)
+   { 
+    Dephone += 10;
+    Cleo += (Cleo * 0.05);
+   }
+  
  cout << "After ";
- cout << year -1 ;
+ cout << year - 1;
  cout << " years";
  cout << ", the value of Cleo's invest can exceed Daphne's ." << endl;
  cout << "The value of Cleo's invest is: " << Cleo << endl;
